nvim: Change integration keybinds
This changes the keybinds from function keys into a chord like how tmux
processes the key sequence from Alacritty. e.g. (C-S-F12 c) will be
forwarded directly to vim now.

Special handling is required to be interoperable between vim and nvim,
and the `Integration` command for binding the key sequence (within
the init.vim script) is the result of that.